+++ app/Views/admin/template.php
@@ -240,6 +240,14 @@
 										</a>
 									</li>
 								</ul>
+								<li class="nav-item">
+								<a href="<?= base_url('admin/blog') ?>" class="nav-link">
+									<i class="nav-icon fas fa-cogs"></i>
+									<p>
+										Configurações
+									</p>
+								</a>
+							</li>
 							</li>
 							<li class="nav-item">
 								<a href="<?= base_url('/login/logout') ?>" class="nav-link">